requirements.txt
requests~=2.31.0
PyJWT~=2.8.0
fastapi~=0.109.0
SQLAlchemy~=2.0.25
pydantic~=2.5.3
uvicorn~=0.26.0
